terraform: ResourceTransformer to ResourceTransformerOld
